requirements.txt
Django==2.1.3
djangorestframework==3.9.0
mysqlclient==1.3.13
pytz==2018.5
